Move ParamDescriptor, IsParameter to core to enable reuse
