+++ dbfit-java/build.gradle
@@ -235,6 +235,7 @@ project('mysql') {
     dependencies {
         compile project(':dbfit-java:core')
         compile publicMysqlCompileDeps
+        testCompile project(':dbfit-java:core').sourceSets.test.output
     }
 
     integrationTest.doFirst {
