+++ dbfit-java/mysql/src/main/java/dbfit/environment/MySqlProcedureParametersParser.java
@@ -1,5 +1,6 @@
 package dbfit.environment;
 
+import dbfit.util.ParamDescriptor;
 import dbfit.util.Direction;
 import static dbfit.util.Direction.*;
